refactor([issue-1025]): source note→frequency from the canonical pitchDetect module
